+++ bomberman/headers/ai.h
@@ -6,7 +6,7 @@
 
 typedef struct S_PathStep
 {
-    double dest_x, dest_y;
+    float dest_x, dest_y;
 
 } PathStep;
 
@@ -47,6 +47,6 @@ void doNextStep( AIModule *module );
 AIModule *createAIModule( Actor *actor, Path path );
 void destroyAIModule( AIModule **module );
 void destroyAIModules( AIModule ** modules, int module_num );
-Actor *createAIActor( Path path, double speed, enum Direction dir, ALLEGRO_BITMAP *bmp );
+Actor *createAIActor( Path path, float speed, enum Direction dir, ALLEGRO_BITMAP *bmp );
 
 #endif
