+++ bomberman/headers/logic.h
@@ -20,10 +20,10 @@ void updatePositionY( Actor *actor );
 
 // ================================== collision_handler.c ==================================
 
-bool isOutOfBounds( Actor *actor, LevelMap *level_map, enum Direction *cdir, int cx, int cy, int cw, int ch );
-void handleOutOfBounds( Actor *actor, LevelMap *level_map, enum Direction cdir, int cx, int cy, int cw, int ch );
-bool isTerrainCollisionX( Actor *actor, LevelMap *level_map, enum Direction *cdir, int cx, int cy, int cw, int ch );
-bool isTerrainCollisionY( Actor *actor, LevelMap *level_map, enum Direction *cdir, int cx, int cy, int cw, int ch );
-void handleTerrainCollision( Actor *actor, enum Direction cdir, int cx, int cy, int cw, int ch );
+bool isOutOfBounds( Actor *actor, LevelMap *level_map, enum Direction *cdir, float cx, float cy, float cw, float ch );
+void handleOutOfBounds( Actor *actor, LevelMap *level_map, enum Direction cdir, float cx, float cy, float cw, float ch );
+bool isTerrainCollisionX( Actor *actor, LevelMap *level_map, enum Direction *cdir, float cx, float cy, float cw, float ch );
+bool isTerrainCollisionY( Actor *actor, LevelMap *level_map, enum Direction *cdir, float cx, float cy, float cw, float ch );
+void handleTerrainCollision( Actor *actor, enum Direction cdir, float cx, float cy, float cw, float ch );
 
 #endif
